Move IsoLayered to isoSpec++.cpp
